move configure-detector to its own subdirectory of tool
