+++ tool/detector/DetectorConfiguration.h
